Add MakeJobStop
Command is used to cancel a running job.

Without an argument, the command will cancel a running MakeJob in one of
two ways:
    1. If the current buffer spawned a MakeJob, that job will be
       stopped.
    2. If the current buffer is displaying MakeJob output, that job will
       stopped.

MakeJobStop also accepts as an argument the name of the MakeJob to stop
(e.g., ":MakeJobStop make"). Command completion can be used to make
finding the name easier.

Additionally do the following refactorings:
1. Add MakeJobStop to help file
2. Update README for MakeJobStop
3. Roll version on script
4. Rename CreateMakeJobWindow to CreateMakeJobBuffer
5. Use <sid> in command definitions according to best practices